Matrix inversion and GETS method
This code is inspired by the GETS method Gaussian Elimination Triangular
Solver method.  I tried to see if a method in MatLab could be adapted to
solve the cash inversion problem. The code needs a lot of work. But here
is the direction I would go with the assignment  if I had a good handle
on functions...I am totally new to programming. Data Toolbox was my
first programming class, ever.